refactor: Refactor index.ts for improved file generation and template selection
