+++ src/get-templates/get-templates.ts
@@ -1,0 +1,5 @@
+import templates from '../templates/acme.json'
+
+export const getTemplates = () => {
+  return templates
+}
