+++ src/make-containing-dir/make-containing-dir.ts
@@ -2,7 +2,7 @@ import { mkdir } from "fs/promises"
 import { ErrorsMessages } from "../errors-messages"
 import { cwd } from "process"
 
-export const makeContainingDir = async (dirName: string): Promise<string | void> => {
+export const makeContainingDir = async (dirName: string): Promise<string> => {
   const dirPath = `${cwd()}/${dirName}`
   try {
     await mkdir(dirPath)
